Fixed some bugs
Envelope does not crash or produce asserts anymore.
Made EnvelopeInterp not constexpr, it was returning NaN for negative
shapes (not sure if this is my fault or the compiler's, since I only
removed constexpr, and did not change any code.)
Added code to break on flop errors, because that's what was causing some
of the bugs.
Fixed crash related to SaveState::getState(index)
Added SimpleKnob::setValue(value) (Where was it?! I thought I added it
before this!)