Added some pitch related stuff.
Added LfoTuner, which takes a number of beats per cycle and turns it
into a frequency that can be put into an oscillator.
Added Pitch Blend, which can perform pitch shifting and sliding on a
frequency.
Added a .jucer file for building the project standalone (which currently
is unusable, as it only shows the basic AtomSynth gui, and does not
receive key presses for playing notes.)