Made a user-workable save / load system.
Fixes #3 (No way to save / load patches)

Added saving / loading of minimum, maximum, and suffixes of automated
knobs, which is useful for knobs that are modified by code. This way,
there does not have to be extra code to set up the range when these
knobs are loaded.
Fixed incorrect version numbers in the save/load code of some atoms.
Removed some duplicate load code from some atoms.
Added a proper properties sidepane for the patch as a whole. From it,
you can change the name of the patch, save it, and see how long it has
been since the last [auto]save.
Simplified the view-switching buttons, instead of having seperate
buttons for sidepane / no sidepane, it is one button which toggles its
icon.
Added SaveManager, which is responsible for saving and loading of
patches as well as keeping track of the name of the current patch.
The Open button will now open a file dialog asking you to pick a patch.
Once a user picks a patch, it will be loaded. This still needs some
polishing, as the main network widget is not repainted and guis from the
old patch are not immediately removed.