Various tweaks.
Fixed bug in BasicOsc when previewing while the phase knob is negative.
Fixed a bug in AtomSlider.
Added a macro for #FF00FF
Added the ability to show messages to the user on the bottom of the
screen.
Save manager now shows a message whenever the patch is saved or another
patch is loaded.