Various polishes / tweaks, it should be ready for a release.
Made it so that AtomManager::setup() will not clear whatever patch is
currently loaded.
A message is now shown to the user if the network has a feedback loop
and can't compile.
The network now recompiles if a connection is disconnected.
Implemented better saving / loading:
The old buttons to save to / load from a file are now disabled by
default.
The data that is usually written to / read from these files is now
written to / read from through the VST API, so that the host can
properly keep track of the status of the synth.
When the old save / load buttons are disabled (as is the default), two
new buttons appear to allow exporting to / importing from the clipboard
as plain text.
SaveManager now has function for importing from / exporting to strings
and byte arrays.
Added a dummy call to Synth::initialize() so that when the gui is loaded
or a preset is imported before AtomSynthAudioProcessor::prepareToPlay(),
they will still work.
The default patch is only loaded if no other patches have been loaded.