Bug fixes
Cleaned up panning in BasicOsc, which might or might not not have fixed
a bug that might or might not have existed.
Applied fmod to internal phase vector of BasicOsc, which fixes #8.
Applied amplitude attenuation after top / bottom remap in BasicOsc,
which fixes #4.
Corrected the panning in the default patch from -0.22 to 0.0.