Maybe fixed #7.
Added code to Synth::initialize() so that when it is called and a patch
is already loaded, the patch is reloaded instead of no action being
taken. I think this will fix #7, because it appears to be caused when
the buffer size is not 256 (the default, there is always one call to
Synth::initialize() with buffer size 256). Previously, the patch would
still have atoms initialized with audio buffers using the old default
size (256). By reloading the patch, it would also recreate the atoms,
and whatever audio buffers they were using, incorporating the new buffer
size, hopefully alleviating the bug.

Heh, there's more characters in this commit message than in the actual
contents of the commit.