Finished the basic workings of the Remap atom.
Fixed some rendering of the grid of plots.
A value can be interpolated from a DrawablePlot now.
Atoms can now set the icons of inputs and outputs later in the code.
Remap now changes the icons of its input and output based on what the
user has selected.
Remap actually works now.