Added CachedFirFilter and other things that make it work.
Previously, it was impractical to automate a filter using an audio
signal. Now, it can be done, with the help of the CachedFirFilter.
FirFilterCache stores multiple different FirCoefficientCache objects,
one for each kind of filter (blackman/lowpass, hamming/lowpass,
blackman/highpass, etc.)
Each FirCoefficientCache has space for 1,029 filter kernels, each kernel
having a different frequency, corresponding to a log base 2 scale
stretching from 16hz to 20khz. When a particular frequency is requested,
the closest 'slot' will be returned.
FirFilterCache can return filter coefficients when given a filter type
and a cutoff frequency. If the requested filter kernel does not exist,
it will be computed.
Because of these additions, when a filter cutoff is automated, it is
very performance-efficient.

Added CachedFirFilter, which instead of using an interal coefficient
array, just uses a pointer to the corresponding kernel in the
coefficient cache.
Modified FirLowpass to use CachedFirFilter instead of FirFilter.