Made SaveStates better.
SaveStates can now import from and export to a more efficient and
lower-level format.
It uses a string as a cointainer for binary data (the string is not
printable, though.)
DrawablePlots can now save from and load to SaveStates.
Automated controls have much more efficient SaveStates.
For min, max, and suffix of an AutomatedControl to be saved, you must
call AutomatedControl::setDynamic(). This is done to save space, since
very few controls actually require this feature.
Updated some atoms with AutomatedControl::setDynamic() and
DrawablePlot::saveSaveState().