Implemented a couple optimizations.
Added looping audio buffer, which is an audio buffer that loops. This is
good for performance of long delay lines, since when you request that
the data be offsetted, only a pointer is offsetted instead of all the
data.
Changed comb atom to use this fancy new buffer.
Added fastPow function from StackOverflow.
Overall, this improved performance on a test patch by ~35% (It used a
comb filter and multi-voice pow wave.)